Fix initial read.
In cecf2b9a430f453848065f3daa521f97c3033dee, cphyc fixed mpi_flag=0 case
but broke the two others. Indeed, the +1 was already given by the main.
Fix this.

Also, the speeds were not on the same line in main_gen, producing huge errors in results.

And end is a language keyword, rather not use it for a var. Change that.