physics: Fix flag_mpi=1
The issue was in front of us, fixing the two FIXME would have fixed the
main issue too: the acceleration returned by compute_force are computed
as the action of the second array over the first one.

For instance, with 4 particles and two MPI :

– first MPI was doing things right ;
– second one was storing −F(3,4 → 1) as F(1,2 → 3) and −F(3,4 → 2) as
F(1,2 → 4).

So, this big bug is solved and will help fixing the FIXME in next
commit.

For now, I’ve added a compute_force_mpi version to fixed this fastly,
but I intend to merge it back into compute_force later.